refactor: Rename commands/server to commands/influxdb_ioxd
